test: add tests for tealium code
